pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deduptickets"
version = "0.1.0"
description = "Ticket Deduplication & Clustering Dashboard"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "DedupTickets Team" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dependencies = [
    "aiohttp==3.11.10",
    "azure-cosmos==4.14.5",
    "azure-functions==1.24.0",
    "azure-identity==1.25.1",
    "fastapi==0.128.0",
    "httpx==0.28.1",
    "openai==1.82.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "python-multipart==0.0.22",
    "uvicorn[standard]==0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "ruff==0.14.14",
    "pre-commit==4.5.1",
    "httpx==0.28.1",
    "bandit[toml]==1.9.3",
]

[project.scripts]
deduptickets = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["backend/src"]

# ============================================================================
# RUFF CONFIGURATION (Constitution Principle VI)
# ============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["backend/src", "backend/tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
    "S",      # flake8-bandit (security)
    "ASYNC",  # flake8-async
]
ignore = [
    "PLR0913",  # Too many arguments - sometimes necessary for Pydantic models
    "PLR2004",  # Magic value used in comparison
]

[tool.ruff.lint.per-file-ignores]
"backend/tests/**/*.py" = [
    "S101",     # Allow assert in tests
    "ARG001",   # Allow unused function arguments (fixtures)
    "PLR2004",  # Allow magic values in tests
]
"backend/src/models/*.py" = [
    "TC001",    # Pydantic needs application imports at runtime for Field defaults
    "TC003",    # datetime/UUID used at runtime in Field defaults
]
"backend/src/schemas/*.py" = [
    "TC001",    # Pydantic needs application imports at runtime for schema resolution
    "TC003",    # datetime/UUID used at runtime in Pydantic field annotations
]
"backend/src/routes/*.py" = [
    "TC001",    # FastAPI needs dependency imports at runtime for DI
    "TC003",    # UUID used at runtime in path parameters
]
"backend/scripts/*.py" = [
    "S311",     # Sample data generator uses pseudo-random intentionally
    "E501",     # Long description strings in data definitions
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ============================================================================
# PYTEST CONFIGURATION (Constitution Principle III)
# ============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
pythonpath = ["backend/src"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=backend/src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests (fast, no I/O)",
    "integration: Integration tests (require Cosmos DB Emulator)",
    "contract: API contract tests",
    "slow: Slow tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# ============================================================================
# COVERAGE CONFIGURATION
# ============================================================================
[tool.coverage.run]
source = ["backend/src"]
branch = true
omit = [
    "*/__init__.py",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true

# ============================================================================
# MYPY CONFIGURATION (Type Checking)
# ============================================================================
[tool.mypy]
python_version = "3.12"
mypy_path = "backend/src"
namespace_packages = true
explicit_package_bases = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "azure.*",
]
ignore_missing_imports = true

# ============================================================================
# BANDIT CONFIGURATION (Security scanning)
# ============================================================================
[tool.bandit]
exclude_dirs = ["backend/tests", ".venv"]
skips = ["B101"]  # Allow assert in production code for Pydantic validators

[dependency-groups]
dev = [
    "faker==40.1.2",
]
